% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fish_families.R
\docType{data}
\name{data_fish_families}
\alias{data_fish_families}
\title{List of families of fish}
\format{
An object of class \code{character} of length 636.
}
\source{
Obtained from the Eschemeyer's Catalog of Fishes and Fishbase.
}
\usage{
data_fish_families
}
\description{
A list of families to filter occurrence data.
}
\keyword{datasets}
