% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_occ_crenicichla.R
\docType{data}
\name{data_occ_crenicichla}
\alias{data_occ_crenicichla}
\title{Occurrence data for Crenicichla in Paraná State, Brazil}
\format{
A data frame with columns such as species, longitude, latitude, and other occurrence details.
}
\source{
Obtained using rFishStatus::rfs_download_occ_data().
}
\usage{
data_occ_crenicichla
}
\description{
A dataset containing occurrence data for Crenicichla species in Paraná State, Brazil.
}
\keyword{datasets}
