% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfs_auto_occ_update.R
\name{rfs_auto_occ_update}
\alias{rfs_auto_occ_update}
\title{Get Synonyms for a Given Species and Update Occurrence Data}
\usage{
rfs_auto_occ_update(ref_df, species, folder)
}
\arguments{
\item{ref_df}{A reference dataset following the rFishStatus template (for this function,
the dataset should be filled with all species from a family).}

\item{species}{A character vector representing the species names to query.}

\item{folder}{A character vector representing the folder path to save the}
}
\value{
A data frame containing updated occurrence data.
}
\description{
This function retrieves synonyms for given species from a
reference list and updates occurrence data for each species (GBIF).
}
\examples{
\dontrun{
result <- rfs_auto_occ_update(ref_df, species, folder)
}
}
