% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfs_download_occ_data.R
\name{rfs_download_occ_data}
\alias{rfs_download_occ_data}
\title{Download occurrence data for a specific area from GBIF}
\usage{
rfs_download_occ_data(
  gbif_user,
  gbif_pwd,
  gbif_email,
  country = "Brazil",
  state = "Paraná",
  taxonkey = 44,
  user_polygon = NULL,
  max_attempts = 10,
  sleep_time = 120,
  reverse_polygon = FALSE,
  folder_name = "occ_data",
  link = NULL
)
}
\arguments{
\item{gbif_user}{GBIF username.}

\item{gbif_pwd}{GBIF password.}

\item{gbif_email}{GBIF email.}

\item{country}{Country name. Default is Brazil.}

\item{state}{State name. Default is Paraná.}

\item{taxonkey}{GBIF taxonKey. Default is Chordata (44).}

\item{user_polygon}{User-defined polygon. Default is NULL.}

\item{max_attempts}{Maximum number of download attempts. Default is 10.}

\item{sleep_time}{Sleep time between download attempts. Default is 120 s.}

\item{reverse_polygon}{Logical. If TRUE, the polygon will be reversed.}

\item{folder_name}{Folder name where the data will be saved. Default is
"occ_data".}

\item{link}{Download link for the occurrence data. Default is NULL.}
}
\value{
Downloaded occurrence data.
}
\description{
This function downloads occurrence data for a specific area from
GBIF. Obs.: This function is unstable and may not work properly due to time
constraints and server limitations while downloading the data. If the first
execution did not work properly, please try to rerun the function using the
same code. Alternatively, you can use the link, showed during function
execution, to download the data manually. IMPORTANT: There is no return
value for this function, data will be saved in the folder specified by the
user.
}
\examples{
\dontrun{
rfs_download_occ_data(gbif_user = "gbif_user",
  gbif_pwd = "gbif_pwd",
  gbif_email = "gbif_email",
  country = "Brazil",
  state = "São Paulo")
}
}
