% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfs_get_species_list.R
\name{rfs_get_species_list}
\alias{rfs_get_species_list}
\title{Get the species list from a dataframe based on rFishStatus template}
\usage{
rfs_get_species_list(df, summary = FALSE)
}
\arguments{
\item{df}{A data frame based on the rFishStatus template (rFishStatus::data_template_ref).}
}
\value{
A dataframe with the species list.
}
\description{
This function uses a dataframe of scientific names from
organized a list of species names with their status.
}
\examples{
df <- rFishStatus::data_template_ref
df_list <- rfs_get_species_list(df)

}
