% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfs_get_synonyms.R
\name{rfs_get_synonyms}
\alias{rfs_get_synonyms}
\title{Get Synonyms for Given Species}
\usage{
rfs_get_synonyms(
  ref_list,
  species,
  all_species = FALSE,
  complete_names = FALSE
)
}
\arguments{
\item{ref_list}{A data frame containing reference species data obtained using
rFishStatus::rfs_get_species_list() function.}

\item{species}{A character vector representing the species names to query.}

\item{all_species}{A logical value indicating whether to return all synonyms
(default is FALSE). If TRUE, returns all synonyms as a vector.}

\item{complete_names}{A logical value indicating whether to return complete
scientific names (default is FALSE) in all species list. If TRUE, returns full scientific names
including author and year.}
}
\value{
A list where each element is named after a valid species and contains
a character vector of its synonyms.
}
\description{
This function retrieves synonyms for given species from a
reference list.
}
\examples{
ref_list <- rFishStatus::rfs_get_species_list(
   rFishStatus::template_ref_data
)
species <- "Crenicichla britskii"
result <- rfs_get_synonyms(ref_list, species)
print(result)
}
