% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfs_get_valid.R
\name{rfs_get_valid}
\alias{rfs_get_valid}
\title{Get Previous Valid Names for Given Species}
\usage{
rfs_get_valid(ref_list, species, all_species = FALSE, complete_names = FALSE)
}
\arguments{
\item{ref_list}{A data frame containing reference species data obtained using
rFishStatus::rfs_get_species_list() function.}

\item{species}{A character vector representing the species names to query.}

\item{all_species}{A logical value indicating whether to return all previous names
(default is FALSE). If TRUE, returns all previous_valid as a vector.}

\item{complete_names}{A logical value indicating whether to return complete
scientific names (default is FALSE) in all species list. If TRUE, returns full scientific names
including author and year.}
}
\value{
A list where each element is named after a valid species and contains
a character vector of its synonyms.
}
\description{
This function retrieves previous valid names for given species from a
reference list.
}
\examples{
ref_list <- rFishStatus::rfs_get_species_list(
   rFishStatus::data_template_ref
)
species <- "Crenicichla britskii"
result <- rfs_get_valid(ref_list, species)
result
}
