% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfs_join_basins.R
\name{rfs_join_basins}
\alias{rfs_join_basins}
\title{Join basins to occurrence dataset}
\usage{
rfs_join_basins(occ_df, shapefile = "br_pr_basins")
}
\arguments{
\item{occ_df}{A data frame of occurrences obtained from
rFishStatus::rfs_update_occ_data().}

\item{shapefile}{A character string specifying the shapefiles contained in the package.
The default is "br_pr_basins". The user can also provide a shapefile directly in sf format.
Obs.: for now, only the shapefile "br_pr_basins" is available directly
from the package. New opstions will be added in the future.}
}
\value{
A data frame of the joined occurrences and basins.
}
\description{
This function joins basins to an occurrence dataset based on
shapefiles. The return is used by rFishStatus::rfs_detect_invasive().
}
\examples{
occ_df <- rFishStatus:::data_occ_update_result
rfs_join_basins(occ_df, "br_pr_basins")
}
