% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfs_mock_function.R
\name{rfs_mock_function}
\alias{rfs_mock_function}
\title{Mock Function}
\usage{
rfs_mock_function(input)
}
\arguments{
\item{input}{A value of any type that will be processed by the function.}
}
\value{
A character string: "This is a mock output".
}
\description{
This function takes an input, prints a message indicating the input received,
and returns a mock output.
}
\examples{
rfs_mock_function("example input")
rfs_mock_function(42)

}
