% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfs_update_occ_data.R
\name{rfs_update_occ_data}
\alias{rfs_update_occ_data}
\title{Update occurrence data}
\usage{
rfs_update_occ_data(
  species_database,
  occurrence_df,
  family_names = "auto",
  folder = "data"
)
}
\arguments{
\item{species_database}{Species database according to rFishStatus template
(see rFishStatus::data_template_ref).}

\item{occurrence_df}{Occurrence dataframe obtained from
rFishStatus::rfs_download_occ_data().}

\item{family_names}{Family names to be used in the filtering
occurrence dataframe. Default is "auto".}

\item{folder}{Folder name where the data will be saved.
Default is "data".}
}
\value{
Occurrence dataframe with updated species names based on the reference
dataset (see columns with "ref" suffix) or FishBase catalog (see columns with
"fb" suffix).
}
\description{
This function checks the occurrence data against a
reference list and update it.
}
\examples{
rfs_update_occ_data(species_database = rFishStatus::data_template_ref,
                    occurrence_df = occ_df)
}
