% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfs_validation.R
\name{rfs_validation}
\alias{rfs_validation}
\title{Helper function to validate species names}
\usage{
rfs_validation(species_database, data, data_type = "spp_list", ref_spp_list)
}
\arguments{
\item{species_database}{Species database according to rFishStatus template
(see rFishStatus::data_template_ref).}

\item{data}{A data frame representing the occurrence data to validate.
Obtained using rFishStatus::rfs_update_occ_data() ("occ_data") or
rFishStatus::rfs_update_spp_list()("spp_list").}

\item{data_type}{A character string representing the type of data to
validate. Should be either "occ_data" or "spp_list". Default is "spp_list".}

\item{ref_spp_list}{A data frame representing the species list to use as
reference. Obtained using rFishStatus::rfs_get_species_list().}
}
\value{
A list with two data frames: attention_df and correct_df.
attention_df contains species that should be manually checked.
}
\description{
This function is used to filter species nomenclature
from rFishStatus::rfs_update_occ_data() that should be manually checked.
}
\examples{
species_database <- rFishStatus::data_template_ref
species_list <- rFishStatus::rfs_get_species_list(species_database)
ref_spp_list <- rFishStatus::rfs_get_species_list(species_database)
attention_list <- rfs_validation(species_database, species_list, "spp_list", ref_spp_list)
}
