% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fish_families.R
\docType{data}
\name{data_fish_families}
\alias{data_fish_families}
\title{List of families of fish}
\format{
A character vector (or data frame, if that's the case) with family names.
}
\source{
Obtained from the Eschemeyer's Catalog of Fishes and Fishbase.
}
\usage{
data_fish_families
}
\description{
A dataset containing a list of fish families to filter occurrence data.
}
\keyword{datasets}
